+++ index.html
@@ -31,7 +31,7 @@
 
     <link rel="stylesheet" href="css/index.css" />
 
-    <script src="js/index.js" defer></script>
+    <script src="js/index.js" type="module"></script>
   </head>
 
   <body>
