+++ js/const.js
@@ -1,0 +1,2 @@
+export const API_URL = 'http://localhost:3024';
+export const PREFIX_PRODUCT = '/api/product';
