+++ js/elements.js
@@ -1,0 +1,19 @@
+export const modalProduct = document.querySelector('.modal_product');
+export const catalogList = document.querySelector('.catalog__list');
+
+export const modalProductTitle = document.querySelector(
+  '.modal-product__title',
+);
+export const modalProductImage = document.querySelector(
+  '.modal-product__image',
+);
+export const modalProductDescription = document.querySelector(
+  '.modal-product__description',
+);
+export const ingredientsList = document.querySelector('.ingredients__list');
+export const ingredientsCalories = document.querySelector(
+  '.ingredients__calories',
+);
+export const modalProductPriceCount = document.querySelector(
+  '.modal-product__price-count',
+);
